added support for skipping sev processing on logs
we really only support severity processing on some kinds of logs,
but we could support timestamps on many more of them. Be explicit
about what we support severity on so that we don't make logs go
away if we ask for them at DEBUG level and they don't have it.

Change-Id: I7460ccfdb4354137988257a9aed319b8dc7297a6