don't doubly escape html files
we want to process console.html files for timestamp links, however
these files are already escaped by jenkins to be valid html, so
our previous assumption that we should always escape lines is
invalid.

based on the filename, don't doubly escape the lines in the file.
Added unit tests to ensure we had expected console.html output
after running it through the filter.

Change-Id: I3a0e6bca782b134bed9f385be056befd3a2f30c7