Support Setting The Path Using an ENV Variable
This commit adds support for setting the root path (in
which to look for logs) using the 'OS_LOGANALYZE_ROOT_PATH'
environment variable.  If no such variable is present,
the old default path ('/srv/static/logs') is used.  The
path may still also be set using the `root_path` parameter
to the `application` function in wsgi.py

This is the first of four commits in a series of commits.

Change-Id: Id83f09c69d83d276c18e14b5e2b4d8ea8640fc2c