add persistence of anchors
It used to be that in jumping between different log levels you
would loose the timestamp anchor that you set. We can fix that
with javascript so that you can set an anchor which is persisted
on the selectors. Then you can jump between different log levels
and not lose your place.

Change-Id: Idf7d204fbff558e3e43e152f601895c5c8567039