make info level not bold
now that we are starting to get debug levels under control, info
is becoming more of a norm. Being bold makes it stand out, but it
also makes it harder to read when all you are looking at is info
lines.

This also creates consistency in that >= WARN is what is bold, and
< is not.

Change-Id: I27bf0ba676c61e8eca84bd83c754a15d0de5e905