fix highlighting on console.html files
the console.html files had improper highlighting because of extra
nesting of pre tags. This was missed the first time around, but
has been driving me nuts for weeks.

The solution is to drop pre tags from files when processing them,
as we're going to wrap it in our own pre tags later. Tests are
updated for the new format.

Change-Id: I5b7f2b0a25ae27a218c51fb73b5fde5055b51578