remove severity support from grenade
An earlier os-loganalyze change added timestamp linking, and
incorrectly also make it support severity levels:

  Ia482f90e1461133ba02f8274577ab1e15932ad3c

However, the severity levels piece was effectively broken by the
fact that grenade.sh.log was a symlink. So it correctly did not
support severity levels.

However, when I932ee1583f04756579abf63c792d336ae937f64d merged
to devstack-gate, grenade.sh.log became a real file. This meant
that the ES indexer stopped seeing any lines because it was filtering
at INFO, which returned nothing.

This change correctly makes it so that grenade.sh.log is ignored
as a file that has severity.

Change-Id: Id65fc616d3b1564a652864573be9844fa0d32ef5