use wsgi environ
We allow you to override the environment of where the logs are
located, however by doing it via os.environ you can't actually
set it in the webserver (you have to set it pre webserver).

Switch to using the wsgi environ which lets you set this via
Apache's SetEnv var, for instance.

Change-Id: I3a0adf91586c03bf238d1674c8b8521a661df3a5