Move all presentation layer to view module
This leverages the prior stack of work which creates a structured
set of line objects and uses them to construct the display. This
is done with a series of views.

This makes for much cleaner display logic, and means some very
weird extra nesting of spans that were done previously is gone.

Tests are updated to support the fact that generator lines now need
not start with </span>.

Change-Id: I7eac7b67e6afcbcf6d46ccf1e2b9682f8c41e56e