Try fetching index.html from swift
OpenStack returns a directory of log files back to gerrit. Swift
has no good way of listing this directory. However we'll upload our
own index.html along with the files so we need to check for that.

If we have 404'd (or failed) on fetching from swift we should try
seeing if an 'index.html' exists.

Change-Id: I1a85a841283ba5f40880356278d5d588139eed7e