add support for source=swift
While we are running both filesystem and swift in parallel, it
would be good to be able to bypass the filesystem to test the
difference between swift and volume performance.

Do this with the addition of a source= cgi parameter. Currently
source=swift is the only thing that will change behavior. Internally
we default it to all.

Refactoring of the cgi param parsing is done to make it cleaner
code and not copy/paste the cgi param parsing a 3rd time.

Change-Id: Ic653bad9a0f74fb5714ff7ccfc7da8d790dcff8b