Try to keep an open connection swift
Reduce load times by removing the need to reauth against keystone
when using a swiftclient. This implementation isn't great as there
is no neat way of keeping a connection object. The generator methods
should probably be refactored into a class at some point, but for now
we'll use a method to establish a new connection or get an existing
one.

Change-Id: Ie9dc339341b13ba672dfcebf9ab5c66f94246ff4