Surpress 404 errors from swift
Because we try swift and then fall back to disk or alternative
path names we get a lot of 404's when looking for a log. These
are currently filling up apache's error.log from our catch-all
exception handler.

We still want to catch all exceptions from attempting swift (as
these will range from 403's through to the swiftclient not existing)
but we want to ignore errors from 404's as they are handled
separately.

Change-Id: I7c03685dfde12a3c09f9a2a9caca811281779875