Only provide html if client supports it
We can specify via config that a file should be served as html but we
should still only do so if the client supports html. Otherwise fallback
on text. This allows us to have a default for log files that is rendered
html in a browser but text when wgetted or viewed in a text editor.

This corrects an issue where we are attempting to index a bunch of html
into elasticsearch because we always get html for the log files even
though the logstash http client doesn't support html.

Change-Id: I35d78498034a9303bd940f57ac34f423673448db