Dedup index entries
When we generate indexes we query disk and swift, if any files have
overlapping paths between them we were generating duplicate index
entries for the same paths. This happened because we used lists to store
the entries which allow for duplicates.

Fix this by using sets until we need to sort (sets are unsorted so we
have to have a list at that point). This will remove any duplicates and
make the index pages more correct.

Change-Id: I6dfa3b30819d6633c3e483d3a386bdce3e26b572