Made test_wsgi somewhat smarter
Now it can distinguish between random strings and actual log messages
(with timestamps and all). The patch also fixes some cases where levels
were incorrectly classified depending on the order of messages parsed.
The fix triggers some renumbering in calculated dicts with number of
matching messages per level.

The test is still a bit dumb because it does not apply elaborate
classification filtering as the actual middleware does, but that's
enough for now.

The patch is needed for the next patch that will add neutron-l3 log
file.

Change-Id: I0232c416a4dbcafb6741a40eeb75a81037cbf0c8