Better supports sev checking
Previously we attempted to check supports sev by running re.search
against the full log path of the log file. This meannt that the re that
roughly said: r'/neutron-' could match job logs like this:

  http://logs.openstack.org/17/553617/16/check/neutron-tempest-plugin-dvr-multinode-scenario/42fb158/job-output.txt.gz

Unfortunately that meant that all neutron logs had to support sev and if
they didn't (which job output doesn't) then we'd serve no logs to the
user if they set a log level.

We address this by switching to matching on the basename of the path and
by switching to re.match so that we only match from the beginning of the
re.

Change-Id: Ia428dc7339f06a15c58001e99ca677cb6059186b