Retire os-loganalyze
This project has been unmaintained for a while. CI jobs don't currently
pass. Zuulv3's dashboard has largely replaced the need for this service
as well as it renders the logs for us instead.

If there is a need for this project elsewhere it can be resurrected and
maintained by others, but for now lets retire it and move on.

Depends-On: https://review.opendev.org/753398
Change-Id: I476b14c7cd9c49270f67897206ec7ad90643703b